FakePath parent also Fake
On FakePath('/arb').parent, delegate to underlying Path, but wrap response in FakePath.
Previously just allowing the FakePath to return .parent as a Path meant that
operations like mkdir going to parent
would not be faked but go to the real fs instead.

Also minor typography fix to config str output - add space to empty known-paths output to standardize
it to look like all the others.