mypy complains about ChainMapPretendDict
mypy has as if a recent release started complaining about
about _ChainMapPretendDict.

given that this class is doing something unusual w multiple
inheritance, and that it is massively flagged as NOT for
external/consumer use, I don't really see the point in
burning time on figuring out the magic incantation to
keep mypy happy, so just add ignore pragma.