tox 4 upgrade
change tox.ini to work with tox 3->4 breaking changes.

specifically `passenv` now takes comma delimited,
rather than space delimited, arguments.

As of tox 4, you HAVE to pass `--` at the cli to
separate options to the test command. In the case
of pypyr, this means instead of:
$ tox ops/build package
you have to run:
$ tox -- ops/build package

Closes #309.